[RNE Rewrite] refactor(ocr): extract pipeline engine into ocrPipeline.ts
Split ocr.ts (1059 lines) so the task file holds only the public API +
createOCR factory:

- Move the tensor-pipeline engine (detectQuads, recognizeQuad,
  recognizeGlyphStrip, readStackedColumn, readBoxVertical), the per-bucket
  builders, and their context/set types into a new internal ocrPipeline.ts
  (imported only by ocr.ts; not re-exported from the package index).
- Extract validateDetectorSchema / buildExtractOpts / disposeDetSets /
  disposeRecSets, removing the duplicated recognizer/detector dispose loops.
- Hoist inline helpers to module scope: pushDetection (ocr.ts), lerp + xSum
  (ocrHelpers.ts).
- Drop the unused DetectContext.format field.

ocr.ts 1059 -> 459 lines. Behavior-preserving; verified on-device (Android):
detector localization + horizontal/vertical recognition unchanged.
